refactor: BinaryStorage 수정
