+++ src/main/java/com/sprint/mission/discodeit/DiscodeitApplication.java
@@ -7,7 +7,9 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 @EnableJpaAuditing
 @SpringBootApplication
 public class DiscodeitApplication {
+
     public static void main(String[] args) {
         SpringApplication.run(DiscodeitApplication.class, args);
     }
+
 }
